Edición de glosarios

+++ AppSapiens/src/main/java/com/mdval/ui/model/cabeceras/DefinicionGlosariosTablaGlosariosCabecera.java
@@ -1,5 +1,6 @@
 package com.mdval.ui.model.cabeceras;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -26,7 +27,7 @@ public class DefinicionGlosariosTablaGlosariosCabecera extends Cabecera {
 		columnIdentifiers.add(literales.getLiteral("tblGlosarios.codUsu"));
 		columnIdentifiers.add(literales.getLiteral("tblGlosarios.fecActu"));
 		
-		columnClasses.add(Integer.class);
+		columnClasses.add(BigDecimal.class);
 		columnClasses.add(String.class);
 		columnClasses.add(Date.class);
 		columnClasses.add(String.class);
